@@ -1,22 +1,26 @@
 {
-  "name": "reactapp",
+  "name": "new",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@fontsource/roboto": "^4.5.8",
     "@material-ui/core": "^4.12.4",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.1",
+    "@material-ui/icons": "^4.11.3",
+    "@mui/material": "^5.11.14",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
-    "framer-motion": "^6.5.1",
-    "install": "^0.13.0",
+    "framer-motion": "^10.9.1",
     "leaflet": "^1.9.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-leaflet": "^3.2.5",
-    "react-router-dom": "^6.9.0",
-    "react-scripts": "^5.0.1",
+    "material-ui-image": "^3.3.2",
+    "mui-image": "^1.0.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.2.1",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
