@@ -1,17 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
-// import App from './App';
-// remember to drop the next line and uncomment the previous line when the right code is available!!!
-import TestApp from './TestApp';
-ReactDOM.render(<TestApp />, document.getElementById('root'));
+import App from './App';
+// import reportWebVitals from './reportWebVitals';
 
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
 
-// const root = ReactDOM.createRoot(document.getElementById('root'));
-// root.render(
-//   <React.StrictMode>
-//     {/* <App />    The main call to kick things off - remember to uncomment it!!! */}
-//     {/*  And remember to take drop the next line!!! */}
-//     <TestApp />
-//   </React.StrictMode>
-// );
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+// reportWebVitals();
